More detailed workshed info added to model train tile (but not CMC tile)
Add workshed prop checking to API.
Make the definition of whether a workshed item is '__iotms_usable' more permissive.
Restrict the nudge to put the cold medicine cabinet in the shed to when it's not already there.
Display different information based on whether the model train set is currently in the workshed (or not).